cohere_invoice_extraction/requirements.txt
streamlit
pdfplumber
cohere
python-dotenv
